+++ nova2d.lua
@@ -1,6 +1,6 @@
 return {
     name    = "Nova2D",
-    version = "0.3.0",
+    version = "0.4.0",
     author  = "Nova2D Contributors",
 
     dependencies = {