+++ src/states/splash.lua
@@ -34,7 +34,7 @@ function State:draw()
         love.graphics.setFont(font)
     end
     local infoFont = love.graphics.setNewFont(14)
-    love.graphics.printf("Nova2D v0.3", 0, 400, 800, "center")
+    love.graphics.printf("Nova2D v0.4", 0, 400, 800, "center")
     love.graphics.setFont(infoFont)
 end